Add mock test for Eagle Cleanup new electronics category
