+++ Dockerfile
@@ -42,6 +42,7 @@ ADD tests/data/www.apache.org.crt /usr/local/share/ca-certificates/
 ADD tests/data/maven.apache.org.crt /usr/local/share/ca-certificates/
 ADD tests/data/www.blender.org.crt /usr/local/share/ca-certificates/
 ADD tests/data/twinery.org.crt /usr/local/share/ca-certificates/
+ADD tests/data/eagle-updates.circuits.io.crt /usr/local/share/ca-certificates/
 ADD docker/create_packages.sh /tmp/
 
 # Refresh the image
