+++ tests/data/eagle-updates.circuits.io.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDpTCCAo2gAwIBAgIJAJu+KDn8qFgVMA0GCSqGSIb3DQEBCwUAMGkxCzAJBgNV
+BAYTAkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEwHwYDVQQKDBhJbnRlcm5ldCBX
+aWRnaXRzIFB0eSBMdGQxIjAgBgNVBAMMGWVhZ2xlLXVwZGF0ZXMuY2lyY3VpdHMu
+aW8wHhcNMTgwMzE2MTQ1NTQ0WhcNMjgwMTIzMTQ1NTQ0WjBpMQswCQYDVQQGEwJB
+VTETMBEGA1UECAwKU29tZS1TdGF0ZTEhMB8GA1UECgwYSW50ZXJuZXQgV2lkZ2l0
+cyBQdHkgTHRkMSIwIAYDVQQDDBllYWdsZS11cGRhdGVzLmNpcmN1aXRzLmlvMIIB
+IjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAwKMatEwQlyWTAhXTbJsUVbz3
+aiWytKBx/13FJs6QGKK8FNmM708dAuoy/oYJintpck3QAQbHwKWszkd4dH04Lmqa
+99m6sSv1b3TR5wk0ubp4rWNSzROz6g4Ob3Pa50jmufE3jb7JxqslGxwEPgwBGfxD
+RJEQD2OoHPLy3ASvUlVgkxRzktkFil6mFL1an2pq/cLxvMMqMysE3hOsCrBNzQSG
+CzpGukh7MAEpeKD11MEtZs42NZHu72iePrT6IIop+A/Qu78nWnfUzeKDyLPhgDe6
+rajMfrpeohW3ihFPxsGpRLWzIUEx2qKBnvs6ZbY7dEAWhRlVVx8A1yG174KrvQID
+AQABo1AwTjAdBgNVHQ4EFgQU4Jk4DqgMeAMPbMUeNTB+aL8tp5wwHwYDVR0jBBgw
+FoAU4Jk4DqgMeAMPbMUeNTB+aL8tp5wwDAYDVR0TBAUwAwEB/zANBgkqhkiG9w0B
+AQsFAAOCAQEAfWe4vZuAhMPopPvM7BZSYY5XFGT9eBTvlMaw28jrIzVJ/njCrGT+
+HkNyfTCx3zPtRa3Sfg/iVOjcqQ96e1bPxsUVlweXa6ZGP3R1PdHvEpgzfXZckHrK
+7QGfDDyoa8+g62GdYiIaOWMUBSY3BFmWyGHIOw1wPSdRdpmtN6DcxNm55D19BwgZ
+9R4UOBGQJKfEKr9nig4KtMfZ+n4Bm45EQq5RlFlHA3wkNTSXcFINQYHuDAmwyYNA
+mBWRDfTJN5dydPNql7yypNl2BHbC5HFRV0WHxz61sxiBWgNCuyWnk85Vx+rGvdxQ
+kGGfgienczq/DFSIWnSf0VB/ywutzpjObg==
+-----END CERTIFICATE-----
