+++ tests/data/eagle-updates.circuits.io.pem
@@ -1,0 +1,50 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDAoxq0TBCXJZMC
+FdNsmxRVvPdqJbK0oHH/XcUmzpAYorwU2YzvTx0C6jL+hgmKe2lyTdABBsfApazO
+R3h0fTguapr32bqxK/VvdNHnCTS5unitY1LNE7PqDg5vc9rnSOa58TeNvsnGqyUb
+HAQ+DAEZ/ENEkRAPY6gc8vLcBK9SVWCTFHOS2QWKXqYUvVqfamr9wvG8wyozKwTe
+E6wKsE3NBIYLOka6SHswASl4oPXUwS1mzjY1ke7vaJ4+tPogiin4D9C7vydad9TN
+4oPIs+GAN7qtqMx+ul6iFbeKEU/GwalEtbMhQTHaooGe+zpltjt0QBaFGVVXHwDX
+IbXvgqu9AgMBAAECggEAakXRBqRsBqaA+oaF9AV962rrZ5DSjNly6nmVuTn9vsmQ
+T0RRfvg7Pg9sAIR7oP1Ch/+rf9V4v6vmsIo3RrSKGPlQ8ukYGIPyzAE7h2TJxKi9
+yJHh7ScwezA0AIJ+wleyieQsTLNzZ3mlZ9ZF9OD0fpxvk9Ufcg39gbL6Di07d4SS
+WVMW7CMRy1NHyWKixNcZV1415HC4zYtujUsXtPlLQxTAdlWdr9som4P9GbA8yCXe
+zXWLFRDYI2HO6IuBtqqGQHKTtdJj+0OPwkia04Y+MM6+CqgHopfDgo0HuvSc3X0e
+dWfHK1RDpAmNvJNnRyMlvzSDBoH/OlF05ZvYAkCx4QKBgQDy0A6wKpR/MvOsxWqN
+rz47XiPhx/MsDR0MsqSIQBTxRFgTmTZCxIjJIQ6Xu+zI35RUxe/MI9RxiYiEPD92
+9FeVHe11Psnhec+2oX2QMQXT6bwNNqwGJhwHZRkgFTUJrHBfq6WsJXjzHhxBgwEz
+UgPiI0QF+UbKB2Tmsq1OmOXA1QKBgQDLGW5tZ4gGXQvjKf3L6kofIvLoxHucegPU
+S0kJ+ZnaA5mqVtb2F7510ZjGiFVHGo5nRbesImFTtKQBm3YZJzmtRCrzvz8GsUiS
+XhkGdFGadq1riMX/PA1tj6ufaHsoQdsZGyc7xqHIh8PqzQ4WjSXKzycHOlpEeYFA
+ENYMVmNzSQKBgFe9uzEJkfOM8+Tgyao9Qx2s3QZ/fG3Up0SpYS8jCO5OJg2e9a6W
+2l7jpSRLcjr8H9uqUwbPg7srWbt/EefoT6XA1MfjUow5ukdif78VF+gFtECKGwt1
+xfceH0Aqqam/ODQ6EHnuaJBhtxAtSMfEwUZ8O950v5HNrjZ3Jo0aTxupAoGAWPjt
+HVRrjI4adQuiOUCL9BP1Ye8hT61MHCZW+t2zfoRxPuQIC4rfTvHhAagb46hIlgk+
+q1PsWyUZa/LUy7QZWkQwpNM1IxNq/WPnhqJSVV6EHkWup1+Bamuz/YNh4ILpHph8
+mX0sxpJjS2tFLq/qzGvvZwUzIigwbQZzHriBnLECgYAPalLTtG6ZLn1ytQFA84lg
+2IHKjFZTVGsT7Q2BBgKkMyMbGtC4Y6S38T+N/TbjKKE3yFvJTa0X75ajTELQ2S4f
+fJU/GxsyyXU54HeDe7QM8ycGmfLTC9eT4lcaOl+EhJBX3XpxZlwVdXIs/aE9hUUl
+JJXzIK6boIGqz0gugxuYOw==
+-----END PRIVATE KEY-----
+-----BEGIN CERTIFICATE-----
+MIIDpTCCAo2gAwIBAgIJAJu+KDn8qFgVMA0GCSqGSIb3DQEBCwUAMGkxCzAJBgNV
+BAYTAkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEwHwYDVQQKDBhJbnRlcm5ldCBX
+aWRnaXRzIFB0eSBMdGQxIjAgBgNVBAMMGWVhZ2xlLXVwZGF0ZXMuY2lyY3VpdHMu
+aW8wHhcNMTgwMzE2MTQ1NTQ0WhcNMjgwMTIzMTQ1NTQ0WjBpMQswCQYDVQQGEwJB
+VTETMBEGA1UECAwKU29tZS1TdGF0ZTEhMB8GA1UECgwYSW50ZXJuZXQgV2lkZ2l0
+cyBQdHkgTHRkMSIwIAYDVQQDDBllYWdsZS11cGRhdGVzLmNpcmN1aXRzLmlvMIIB
+IjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAwKMatEwQlyWTAhXTbJsUVbz3
+aiWytKBx/13FJs6QGKK8FNmM708dAuoy/oYJintpck3QAQbHwKWszkd4dH04Lmqa
+99m6sSv1b3TR5wk0ubp4rWNSzROz6g4Ob3Pa50jmufE3jb7JxqslGxwEPgwBGfxD
+RJEQD2OoHPLy3ASvUlVgkxRzktkFil6mFL1an2pq/cLxvMMqMysE3hOsCrBNzQSG
+CzpGukh7MAEpeKD11MEtZs42NZHu72iePrT6IIop+A/Qu78nWnfUzeKDyLPhgDe6
+rajMfrpeohW3ihFPxsGpRLWzIUEx2qKBnvs6ZbY7dEAWhRlVVx8A1yG174KrvQID
+AQABo1AwTjAdBgNVHQ4EFgQU4Jk4DqgMeAMPbMUeNTB+aL8tp5wwHwYDVR0jBBgw
+FoAU4Jk4DqgMeAMPbMUeNTB+aL8tp5wwDAYDVR0TBAUwAwEB/zANBgkqhkiG9w0B
+AQsFAAOCAQEAfWe4vZuAhMPopPvM7BZSYY5XFGT9eBTvlMaw28jrIzVJ/njCrGT+
+HkNyfTCx3zPtRa3Sfg/iVOjcqQ96e1bPxsUVlweXa6ZGP3R1PdHvEpgzfXZckHrK
+7QGfDDyoa8+g62GdYiIaOWMUBSY3BFmWyGHIOw1wPSdRdpmtN6DcxNm55D19BwgZ
+9R4UOBGQJKfEKr9nig4KtMfZ+n4Bm45EQq5RlFlHA3wkNTSXcFINQYHuDAmwyYNA
+mBWRDfTJN5dydPNql7yypNl2BHbC5HFRV0WHxz61sxiBWgNCuyWnk85Vx+rGvdxQ
+kGGfgienczq/DFSIWnSf0VB/ywutzpjObg==
+-----END CERTIFICATE-----
