+++ tests/data/server-content/eagle-updates.circuits.io/downloads/latest.html
@@ -1,0 +1,72 @@
+'
+<html>
+<head>
+    <title>Autodesk EAGLE 8.7.0 Download</title>
+    <style>
+        @font-face {
+                font-family: 'Artifakt Element';
+                src: url('https://eagle-updates.circuits.io/8_0_0_update/Artifakt+Element+Regular.woff') format('woff');
+                font-weight: normal;
+                font-style: normal;
+        }
+	body{
+		font-family: 'Artifakt Element';
+		margin: 0;
+	}
+	.heading{
+		background: #323232;
+		padding: 20px 30px 10px 30px;
+		color: white;
+	}
+	.heading h3{
+		margin-bottom: 0;
+		font-weight: normal;
+	}
+	.footer{
+		padding-top: 10px;
+		border-top: 1px solid #F2F2F2;
+		text-align: right;
+		padding-right: 30px;
+	}
+	li{
+		font-size: 10px;
+	}
+	li span{
+		font-size: 12px;
+	}
+        li span a {
+                padding: 0;
+        }
+	a{
+		padding: 10px 20px;
+		border-radius: 5px;
+		border: none;
+		font-size: 14px;
+		font-family: 'Artifakt Element';
+		margin: 0px 3px;
+                text-decoration: none;
+	}
+
+
+    </style>
+
+
+</head>
+
+<body>
+<div class="heading">
+    <img src="https://eagle-updates.circuits.io/8_0_0_update/eagle_icon.svg">
+    <h3>EAGLE 8.7.0 Download</h3>
+</div>
+
+<div class="content">
+    <ul style="font-weight: 300;">
+        <li><span><a href="https://eagle-updates.circuits.io/downloads/8_7_0/Autodesk_EAGLE_8.7.0_English_Win_64bit.exe">EAGLE 8.7.0 for Windows 64 bit</a></span></li>
+        <li><span><a href="https://eagle-updates.circuits.io/downloads/8_7_0/Autodesk_EAGLE_8.7.0_English_Mac_64bit.pkg">EAGLE 8.7.0 for macOS</a></span></li>
+        <li><span><a href="https://eagle-updates.circuits.io/downloads/mock/Autodesk_EAGLE_mock_English_Linux_64bit.tar.gz">EAGLE 8.7.0 for Linux</a></span></li>
+    </ul>
+</div>
+
+
+</body>
+</html>
